feat: update readme and tests
