+++ lib/index.js
@@ -17,12 +17,12 @@ function createTestProxy(tapTest) {
 function decorateTestMethod(tap) {
   const originFn = tap.test
   tap.test = async (name, fn) => {
-    const wrapperFn = async (testObject) => {
-      decorateTestMethod(testObject)
-      const proxyTest = createTestProxy(testObject)
-      return await fn.call(this, proxyTest)
+    const wrapperFn = (tap) => {
+      decorateTestMethod(tap)
+      const proxyTest = createTestProxy(tap)
+      return fn.call(tap, proxyTest)
     }
-    return originFn.call(this, name, wrapperFn)
+    return originFn.call(tap, name, wrapperFn)
   }
 }
