+++ lib/testsMethod.js
@@ -14,5 +14,4 @@ module.exports = function testsAssertion(name, dataSource, callback) {
       callback(t, ...source[i], i)
     )
   }
-  return this.pass()
 }
